requirements.txt
dnspython
simplejson
